Remove unused functions and refactor the TypedApiContext module for better organization.
