+++ generation/src/main/kotlin/io/github/nomisrev/openapi/Naming.kt
@@ -114,7 +114,7 @@ private class Nam(private val `package`: String) : Naming {
 
   override fun toFunName(route: Route): String =
     when (val operationId = route.operationId) {
-      null -> "${route.path.segments().last()}${route.method.name()}"
+      null -> "${route.path.pathSegments().last()}${route.method.name()}"
       else -> operationId
     }.toCamelCase()
